+++ lib/markawesome/transformer.rb
@@ -12,6 +12,7 @@ module Markawesome
 
       content = LayoutTransformer.transform(content)
       content = PopoverTransformer.transform(content)
+      content = TooltipTransformer.transform(content)
       content = BadgeTransformer.transform(content)
       content = ButtonTransformer.transform(content)
       content = CalloutTransformer.transform(content)